// reactcodewebapp-main/src/services/api.d.ts
export interface TripPoint {
    lat: number;
    lng: number;
    speed: number;
    timestamp: string;
    address: string;
}